Fix conversion warnings in client code
Issue #1653.

+++ client/client_shared.h
@@ -50,7 +50,7 @@ struct mosq_config {
 	int pub_mode; /* pub, rr */
 	char *file_input; /* pub, rr */
 	char *message; /* pub, rr */
-	long msglen; /* pub, rr */
+	int msglen; /* pub, rr */
 	char *topic; /* pub, rr */
 	char *bind_address;
 	int repeat_count; /* pub */
@@ -65,7 +65,7 @@ struct mosq_config {
 	char *password;
 	char *will_topic;
 	char *will_payload;
-	long will_payloadlen;
+	int will_payloadlen;
 	int will_qos;
 	bool will_retain;
 #ifdef WITH_TLS
@@ -100,7 +100,7 @@ struct mosq_config {
 	bool eol; /* sub */
 	int msg_count; /* sub */
 	char *format; /* sub */
-	int timeout; /* sub */
+	unsigned int timeout; /* sub */
 	int sub_opts; /* sub */
 	long session_expiry_interval;
 #ifdef WITH_SOCKS
